refactor: separate dateRange, widgetLocation, consoleFilter from base widget state
