+++ apps/web/src/services/dashboards/widgets/_hooks/use-widget/use-widget-date-range.ts
@@ -1,0 +1,22 @@
+
+import type { ComputedRef, UnwrapRef } from 'vue';
+import { computed, isRef } from 'vue';
+
+import dayjs from 'dayjs';
+
+import type { DateRange } from '@/services/dashboards/config';
+import { GRANULARITY } from '@/services/dashboards/widgets/_configs/config';
+import type { WidgetProps } from '@/services/dashboards/widgets/_configs/config';
+import type { BaseWidgetState } from '@/services/dashboards/widgets/_hooks/use-widget/use-base-widget-state';
+import type { AdditionalState } from '@/services/dashboards/widgets/_hooks/use-widget/use-widget';
+
+export const useWidgetDateRange = (props: WidgetProps, widgetState: UnwrapRef<BaseWidgetState>, overrides: AdditionalState = {}): ComputedRef<DateRange> => {
+    const dateRange = computed<DateRange>(() => {
+        if (overrides.dateRange) return isRef(overrides.dateRange) ? overrides.dateRange.value : overrides.dateRange;
+        const dateRangeFormat = widgetState.options.granularity === GRANULARITY.YEARLY ? 'YYYY' : 'YYYY-MM';
+        const end = dayjs.utc(widgetState.settings?.date_range?.end).format(dateRangeFormat);
+        const start = dayjs.utc(widgetState.settings?.date_range?.start).format(dateRangeFormat);
+        return { start, end };
+    });
+    return dateRange;
+};
